requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
facenet-pytorch==2.6.0
fastapi==0.117.1
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.4
keras==2.15.0
libclang==18.1.1
Markdown==3.9
MarkupSafe==2.1.5
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.3.1
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==10.2.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-multipart==0.0.20
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
Werkzeug==3.1.3
wrapt==1.14.2
